1. 消费如果失败，退出对话框并回调onPaymentCancel 2. 调整菜品主类、菜品子类、菜品的显示高度和宽度 3. 记录明细中的最后一个方式，修改为状态

+++ app/src/main/java/com/rankway/controller/activity/project/DeskPosLoginActivity.java
@@ -98,7 +98,7 @@ public class DeskPosLoginActivity
         } catch (Exception e) {
             e.printStackTrace();
         }
-        tvAppVersion.setText(String.format("程序版本：1.0." + versionCode));
+        tvAppVersion.setText("程序版本：1.0." + versionCode);
 
         //  上次缓存的用户名
         String str = SpManager.getIntance().getSpString(AppIntentString.LAST_LOGIN_USER);
@@ -160,7 +160,8 @@ public class DeskPosLoginActivity
 
     /***
      * 验证操作员密码
-     * @return
+     * @return  true        验证成功
+     *          false       验证失败
      */
     private boolean verifyUserCode() {
         String struserid = etUserCode.getText().toString().trim();
@@ -254,8 +255,10 @@ public class DeskPosLoginActivity
             } else {
                 sendProccessMessage("同步 操作员信息 成功");
                 Log.d(TAG, "操作员个数：" + listUserInfo.size());
-                DBManager.getInstance().getUserInfoEntityDao().deleteAll();
-                DBManager.getInstance().getUserInfoEntityDao().saveInTx(listUserInfo);
+                if(listUserInfo.size()>0) {
+                    DBManager.getInstance().getUserInfoEntityDao().deleteAll();
+                    DBManager.getInstance().getUserInfoEntityDao().saveInTx(listUserInfo);
+                }
                 n++;
             }
